Switch to long hash
This will make the URL seamlessly render as a code block in the context of GitHub markdown.

+++ src/main.js
@@ -81,7 +81,7 @@ module.exports = {
       url: config[`remote "${remoteName}"`].url, // An URL to git repository itself.
       githubUrl: githubUrlFromGit(config[`remote "${remoteName}"`].url, gitUrl && {extraBaseUrls: [gitUrl]}), // An URL to the GitHub page for given repository.
       // Include hash only on demand as it might be a costy operation.
-      hash: includeHash ? gitRevSync.short(cwd) : null
+      hash: includeHash ? gitRevSync.long(cwd) : null
     }
   }
 }
